Backend/requirements.txt
fastapi
uvicorn[standard]
pypdf
mysql-connector-python

